# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/doctors": {
      "get": {
        "description": "Returns doctors with optional status, categoryId, search filters and pagination.",
        "operationId": "AdminController_listDoctors",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "DEACTIVATED"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter to doctors assigned to this categoryId.",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Substring match on name or category name",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Doctors and pagination metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "doctors": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "List doctors",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminController_createDoctor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Doctor created."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Create a new doctor record",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/doctors/{id}": {
      "get": {
        "operationId": "AdminController_getDoctor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Doctor id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Doctor not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Get a doctor by id",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_updateDoctor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Doctor id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Doctor not found."
          },
          "409": {
            "description": "No fields to update."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Update a doctor (partial)",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteDoctor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Doctor id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Doctor deleted."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Doctor not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Hard-delete a doctor",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/doctors/{id}/deactivate": {
      "patch": {
        "operationId": "AdminController_deactivateDoctor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Doctor id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Doctor not found."
          },
          "409": {
            "description": "Doctor is already deactivated."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Soft-deactivate a doctor",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Get a user by id",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminController_changeUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleChangeDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Invalid role value."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Demotion rejected — last active admin."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Promote or demote a user",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}/deactivate": {
      "patch": {
        "operationId": "AdminController_deactivateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Deactivate a user account",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/doctors": {
      "get": {
        "description": "Anonymous-accessible listing of doctors. Optional filters: categoryId (FK equality, requires the category to be ACTIVE) and search (case-insensitive substring on name + category.name). Pagination via page (default 1) and pageSize (default 20, max 100).",
        "operationId": "DoctorsController_listDoctors",
        "parameters": [
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter to doctors in this category (and only ACTIVE categories).",
            "schema": {
              "maxLength": 64,
              "example": "seed_cardiology",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on doctor name OR category name.",
            "schema": {
              "maxLength": 120,
              "example": "Smith",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of doctors.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "doctors": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "List ACTIVE doctors (public)",
        "tags": [
          "doctors"
        ]
      }
    },
    "/api/doctors/{id}": {
      "get": {
        "description": "Returns the full public record of a single ACTIVE doctor whose category is also ACTIVE. Returns 404 otherwise.",
        "operationId": "DoctorsController_getDoctor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Doctor id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "doctor": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Get a public doctor profile",
        "tags": [
          "doctors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vezeeta Backend API",
    "description": "REST API for the Vezeeta backend.\n\n## Auth\n\nAuth is handled by [Better Auth](https://www.better-auth.com/) and exposed under `/api/auth/*`.\nAfter `sign-in/email` or `sign-in/phone-number`, the server sets an HTTP-only session cookie.\nThe frontend must send the cookie on every subsequent request (use `credentials: \"include\"` in fetch / `withCredentials: true` in axios).\n\nFor the social OAuth flow, the browser hits `GET /api/auth/oauth/start?provider=google` (or `facebook`) — a Better Auth standard route — and follows the 302 redirect.\n\nUse the **Authorize** button below to paste the session cookie value if you want to call protected endpoints from Swagger UI.\n\n## Tags\n\n- `auth` — current-session, social linking, health\n- `doctors` — public doctor catalog\n- `categories` — public category dropdown\n- `slots` — public slot picker for a doctor\n- `doctor-services` — per-doctor service catalog (admin)\n- `appointments` — patient booking flow\n- `reviews` — patient reviews + public doctor rating\n- `medical-records` — patient-facing read of their records\n- `notifications` — in-app inbox\n- `admin` — admin-only management (requires the `admin` role)",
    "version": "0.0.1",
    "contact": {},
    "x-generated-at": "2026-07-22T17:35:45.274Z"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vezeta.session_token",
        "description": "Better Auth session cookie set by /api/auth/* (prefix configured in src/auth/auth.ts)."
      }
    },
    "schemas": {
...
      "CreateDoctorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Doctor's full display name.",
            "minLength": 2,
            "maxLength": 120,
            "example": "Dr. Jane Smith"
          },
          "categoryId": {
            "type": "string",
            "description": "Id of the ACTIVE category this doctor belongs to.",
            "example": "seed_cardiology"
          },
          "bio": {
            "type": "string",
            "description": "Short biography / about section.",
            "maxLength": 2000
          },
          "imageUrl": {
            "type": "string",
            "description": "Profile photo URL. Must be a valid https:// URL.",
            "maxLength": 2048,
            "example": "https://cdn.example.com/jane.jpg"
          }
        },
        "required": [
          "name",
          "categoryId"
        ]
      },
      "UpdateDoctorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "categoryId": {
            "type": "string",
            "description": "Id of an ACTIVE category to reassign the doctor to.",
            "maxLength": 64
          },
          "bio": {
            "type": "string",
            "maxLength": 2000
          },
          "imageUrl": {
            "type": "string",
            "maxLength": 2048
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DEACTIVATED"
            ]
          }
        }
      },
      "RoleChangeDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}